Added a DEFINE case to test a hypothesis.

+++ ManuvrOS/EventManager.cpp
@@ -321,7 +321,7 @@ int8_t EventManager::procIdleFlags() {
   uint8_t activity_count    = 0;     // Incremented whenever a subscriber reacts to an event.
 
   /* As long as we have an open event and we aren't yet at our proc ceiling... */
-  while (event_queue.hasNext() && (EVENT_MANAGER_MAX_EVENTS_PER_LOOP > return_value)) {
+  while (event_queue.hasNext() && ((int8_t) EVENT_MANAGER_MAX_EVENTS_PER_LOOP > return_value)) {
     active_event = event_queue.dequeue();       // Grab the Event and remove it in the same call.
     msg_code_local = active_event->event_code;  // This gets used after the life of the event.
